CoinAPI refactoring includes new parameters in ToolBox
